Continued: Remove support for models with separate encoder and decoder
* Renamed train_byt5model.py to train_model.py
* Include the updates in dataset_stats.py and tests for train_model
